requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
APScheduler>=3.10
playwright>=1.55
httpx>=0.27
pydantic>=2.8
pendulum>=3.0
loguru>=0.7
